Moving FunctionPipeline to new pipelines/. Adding custom to supported tasks
