+++ tests/test_pipeline.py
@@ -14,7 +14,9 @@
 
 from typing import OrderedDict
 
-import numpy as np
+import numpy
+from PIL import Image
+from torchvision import transforms
 
 import pytest
 from deepsparse.image_classification import (
@@ -23,13 +25,9 @@ from deepsparse.image_classification import (
     ImageClassificationInput,
     ImageClassificationOutput,
 )
-from deepsparse.pipeline import FunctionPipeline
+from deepsparse.pipeline import Pipeline
 from sparsezoo import Zoo
 from sparsezoo.utils import load_numpy_list
-
-
-from PIL import Image  # isort:skip
-from torchvision import transforms  # isort:skip
 
 
 @pytest.mark.parametrize(
@@ -57,16 +55,16 @@ def test_function_pipeline_as_image_classifier(zoo_stub, image_size):
         assert len(input.images) == 1
         assert isinstance(input, ImageClassificationInput)
         assert isinstance(input.images, list)
-        assert all(isinstance(i, np.ndarray) for i in input.images)
+        assert all(isinstance(i, numpy.ndarray) for i in input.images)
         return [
             standard_imagenet_transforms(Image.fromarray(img)).unsqueeze(0).numpy()
             for img in input.images
         ]
 
-    def postprocess(outputs):
-        # assert len(outputs) == 1
+    def postprocess(outputs, **kwargs):
+        assert len(outputs) == 2  # NOTE: logits & softmax for this model
         labels, label_scores = [], []
-        for scores in outputs[0]:
+        for scores in outputs[1]:
             lbl = scores.argmax(-1)
             labels.append(lbl)
             label_scores.append(scores[lbl])
@@ -76,7 +74,8 @@ def test_function_pipeline_as_image_classifier(zoo_stub, image_size):
             labels=labels,
         )
 
-    pipeline = FunctionPipeline(
+    pipeline = Pipeline.create(
+        "custom",
         zoo_stub,
         input_schema=ImageClassificationInput,
         output_schema=ImageClassificationOutput,
@@ -90,7 +89,7 @@ def test_function_pipeline_as_image_classifier(zoo_stub, image_size):
     assert isinstance(sample, OrderedDict)
     assert len(sample) == 1
     image_raw = list(sample.values())[0]
-    assert isinstance(image_raw, np.ndarray)
+    assert isinstance(image_raw, numpy.ndarray)
     input = ImageClassificationInput(images=[image_raw])
     output = pipeline(input)
     assert isinstance(output, ImageClassificationOutput)
